move to `addons/trigger_bomb_reset`
