+++ regamedll/dlls/triggers.cpp
@@ -1814,25 +1814,6 @@ void CTriggerTeleport::Spawn()
 	SetTouch(&CTriggerTeleport::TeleportTouch);
 }
 
-LINK_ENTITY_TO_CLASS(trigger_bomb_reset, CTriggerBombReset, CCSTriggerBombReset)
-
-void CTriggerBombReset::Spawn()
-{
-	InitTrigger();
-	SetTouch(&CTriggerBombReset::Touch);
-}
-
-void CTriggerBombReset::Touch(CBaseEntity *pOther)
-{
-	CWeaponBox *pWeaponBox = dynamic_cast<CWeaponBox *>(pOther);
-
-	if (pWeaponBox && pWeaponBox->m_bIsBomb)
-	{
-		// If the bomb touches this trigger, tell it to reset to its last known valid position.
-		pWeaponBox->ResetToLastValidPlayerHeldPosition();
-	}
-}
-
 LINK_ENTITY_TO_CLASS(info_teleport_destination, CPointEntity, CCSPointEntity)
 LINK_ENTITY_TO_CLASS(func_buyzone, CBuyZone, CCSBuyZone)
