+++ regamedll/msvc/ReGameDLL.vcxproj.filters
@@ -1,4 +1,4 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
     <Filter Include="engine">
@@ -549,6 +549,9 @@
     <ClCompile Include="..\dlls\addons\point_command.cpp">
       <Filter>dlls\addons</Filter>
     </ClCompile>
+    <ClCompile Include="..\dlls\addons\trigger_bomb_reset.cpp">
+      <Filter>dlls\addons</Filter>
+    </ClCompile>
     <ClCompile Include="..\public\FileSystem.cpp">
       <Filter>public</Filter>
     </ClCompile>
@@ -1047,6 +1050,9 @@
       <Filter>dlls\addons</Filter>
     </ClInclude>
     <ClInclude Include="..\dlls\addons\point_command.h">
+      <Filter>dlls\addons</Filter>
+    </ClInclude>
+    <ClInclude Include="..\dlls\addons\trigger_bomb_reset.h">
       <Filter>dlls\addons</Filter>
     </ClInclude>
   </ItemGroup>
