+++ regamedll/regamedll/dlls.h
@@ -134,6 +134,7 @@ using FloatRef = float;
 #include "addons/trigger_setorigin.h"
 #include "addons/trigger_random.h"
 #include "addons/point_command.h"
+#include "addons/trigger_bomb_reset.h"
 
 // Tutor
 #include "tutor.h"
